Revert "Update route.ts"
