--- middleware.ts
@@ -1,30 +1,0 @@
-// middleware.ts
-import { NextResponse } from "next/server";
-import type { NextRequest } from "next/server";
-
-const publicRoutes = ["/auth/signin", "/auth/signup", "/api/public-route"];
-
-export function middleware(request: NextRequest) {
-  const { pathname } = request.nextUrl;
-  const token = request.cookies.get("next-auth.session-token")?.value;
-
-  if (publicRoutes.includes(pathname) || pathname.startsWith("/_next/")) {
-    return NextResponse.next();
-  }
-
-  if (!token) {
-    const signInUrl = new URL("/auth/signin", request.url);
-    signInUrl.searchParams.set("redirect", pathname);
-    return NextResponse.redirect(signInUrl);
-  }
-
-  if (pathname === "/") {
-    return NextResponse.redirect(new URL("/dashboard", request.url));
-  }
-
-  return NextResponse.next();
-}
-
-export const config = {
-  matcher: ["/((?!_next/static|favicon.ico|assets).*)"],
-};
