--- next-auth.d.ts
@@ -1,11 +1,0 @@
-import NextAuth from "next-auth";
-
-declare module "next-auth" {
-  interface Session {
-    id: string;
-  }
-
-  interface JWT {
-    id: string;
-  }
-}
